daemon: added configuration directive for setting resolver ruleset.
Added a new configuration directive that can be used to set various
configuration settings. Currently it can only be used to set the
resolver ruleset path like this

set resolver-ruleset '/path/to/resolver/ruleset-file'

It can be generalized with a few minor changes to handle other
configuration settings if/when needed.