common/msg: added mrp_msg_get for fetching multiple message fields.
You can use mrp_msg_get to fetch multiple fields from a message with
a single function call. Moreover, the implementation will do its best
to avoid scanning the message more than once. IOW, mrp_msg_get can be
used to fetch multiple fields from a message efficiently if the fields
to fetch are specified in the same order (with potential fields missing
in between) as they are present in the actual message.