common: a bit better superloop abstraction with improvement cleanup
Cleaned up a little bit the superloop abstraction with repect to
cleanup/unregistration. Now the superloop glue code provides a
cleanup handler which is invoked by the mainloop during unregistration.
The dangerous bits of passing the opaque glue data between the mainloop
and the superloop glue via the user are now gone.

Updated the pulse glue accordingly.