native_resource: Implement a lot of things
We now run the Murphy main loop in its own thread, so that it
does not block our own stuff. Also we actually create a resource
set and try to acquire a resource.

Currently fails with "E: ignoring malformed resource event", but
this might be because of the initial messages from Murphy. PEBKAC
of course possible, as always